move common SpringBoot beans to a separate config
